fix: fetch based on total
